most tests are very obsolete; sweeping them under the rug test project now builds again, with serialization support fixing some json formatting & support for empty arrays & objects adding harmonizer setting serialization

+++ src/clarinoid/basic/Enum.hpp
@@ -48,7 +48,7 @@ struct GenericEnumInfo : IList
 
     int List_GetItemCount() const
     {
-        return mItemCount;
+        return (int)mItemCount;
     }
 
     String List_GetItemDisplayName(int n) const
